Update README.md
When running the barman-exporter via a cronjob, we may need to specify the exactly path of the execution, otherwise the job will run but unsuccessfully. 

Also, in my experience I found it easier to run it from the `root` user directly, so it's not necessary to add `root` before the command.